requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
argparse==1.4.0
    # via -r requirements.in
backoff==2.2.1
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
crossref==0.1.2
    # via -r requirements.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.179.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via httplib2
pypdf2==3.0.1
    # via -r requirements.in
requests==2.32.4
    # via google-api-core
rsa==4.9.1
    # via google-auth
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
